TextArea interfered with Flicksble. Fixed this.
Due to TextArea component having its own Flickable to scroll text inside,
it interfered with MainPage's Flickable and it was really hard to scroll
the page if TextAreas fill most of it. Overcame this issue by moving
translation TextArea into its own page while MainPage contains Label and
opens the translation page when it's clicked. This approach still allows
to select and copy the translation text.

This change improves #1 - Translation doesn't fit on the screen when
translating large texts.