SYM: Workaround a problem with VKB
When having VKB open, switching to application with VKB open and
returning back, `inputContext.visible` is set to false while VKB is
still visible. This seems to be a bug in Qt or even lower.

The workaround assumes that when Symbian reports that application window
is partially visible and text input area is focused, then this is VKB
that partially covers the window.